add parsing support for sympy Piecewise, fixes #1591 (#1594)
* add parsing support for sympy Piecewise, fixes #1591

* cleanup

* move funs to import utils
