+++ .gitignore
@@ -154,6 +154,7 @@ python/examples/example_presimulation/model_presimulation_re/*
 python/examples/example_constant_species/model_constant_species_reduced/*
 python/examples/example_constant_species/model_constant_species/*
 python/tests/sbml_test_models/*
+python/tests/piecewise_test/*
 
 python/sdist/amici.egg-info/*
 python/sdist/amici/version.txt
